pyproject.toml
[project]
name = "freequeensy-int20h-ai"
version = "0.1.0"
description = "Dataset generator and quality analyzer for customer support chat dialogues using OpenAI API"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.24.0",
    "instructor>=1.14.5",
    "python-dotenv>=1.2.1",
]
